requirements.txt
Google-Search-API==1.1.14
python-docx==0.8.10
sumy==0.7.0